refactor(agent): extract dispatch domain from loop
